Rename lco_ingester to ocs_ingester
