+++ listener.py
@@ -6,7 +6,7 @@ from kombu.mixins import ConsumerMixin
 from kombu import Connection, Queue, Exchange
 
 import tasks
-from lco_ingester.settings import settings
+from ocs_ingester.settings import settings
 
 logger = logging.getLogger('lco_ingester')
